refactor(#271): split useResourceProfile into focused sub-modules
Behaviour-preserving refactor. Extracts the monolithic hook into six
focused modules with clear ownership boundaries:

- useResourceProfileData: data fetching queries
- useResourceProfileMutations: resource type, named resource, overhead CRUD
- useAllocationEditing: allocation mode editing state + mutations
- useCommercialData: commercial queries, chart data, filtered rows
- useCommercialMutations: discount/tax/rate card mutations
- useResourceProfileExport: CSV export helpers

useResourceProfile remains as a facade composing all sub-modules
with the identical return shape — zero UI/consumer changes.

All mutations use central projectInvalidation helpers from #267.
No server API changes. No calculation semantics changed.
